Change to a semi-strongly-typed Error type and a single-level result.
This changes the return type for routes from
    Result<Result<T, E>, failure::Error>
to
    Result<T, dropbox_sdk::Error<E>>
by introducing a new error type specific to this crate which holds the
strongly-typed error as one of its variants, but also can contain any
other error in another of its variants.

This should make error handling more idiomatic and ergonomic.