failure changes: don't wrap RequestError in Context so much
The context on RequestError is not helpful; it gets in the way of
downcasting and doesn't add any details.

Also remove Error::RequestError and just put those in Error::Other.
Users are probably going to be downcasting anyway, so this just makes it
simpler: one less case to have to write out.