flatten result type
This changes routes from returning:
    Result<Result<T, E>, dropbox_sdk::Error>
to return:
    Result<T, dropbox_sdk::Error<E>>

The Error type is now type-parametrized on the error type defined by
each route, which it holds as Error::API, and an Error::Other variant
which holds a GeneralError that is not type-parametrized for other
errors not specific to a route.

This is a super breaking change, but removes the biggest ugliness with
the API, hopefully in a relatively ergonomic way.