async proof-of-concept
This still uses the non-async Hyper 0.10, so it doesn't actually result
in any performance changes, but the functions are all now `async` and
this proves that it can be done reasonably.

Main ugliness right now is the fact that the `client` parameter needs to
include a type parameter for the type of future returned by the HTTP
client. Callers shouldn't have to write this, because they don't care,
but the lack of `async fn` in traits makes it necessary.