upgrade to hyper 0.13 -- part 3 (fix large-file-upload example)
It's unfortunate that the upload endpoints now have to take ownership of
the buffer passed in... Maybe that type should be changed from `Vec<u8>`
to `Bytes` so that copies can be avoided in most cases.

Or maybe it should be changed even further to allow passing in a `Read`
or a `Stream` (of u8, or Bytes, or...) or other types.

Orrrrrr maybe there could be multiple functions, like say
`files::upload_vec`, `files::upload_stream`, `files::upload_reader`, and
so on. Hmm.