Take request body as stream
Stream of Result<Bytes, AnyError> for now. Not sure if Bytes is the best
way for callers to pass in data, it's kind of awkward to use, but it's
what Hyper wants, so it makes implementing HyperClient easy :P

Also this change removes the explicit lifetime on all parameters and
return type in request(); makes the body stream `'static' and takes the
argument by value instead of by reference.