remove generic lifetimes from client trait
This makes me a little sad.

Ideally, both the request body and response body could be borrowed, as
long as the request body outlives the request lifetime, and the response
body is bounded by the request lifetime. But Hyper uses Bytes, which
imposes a 'static bound on both of these.

I experimented with making HttpClient generic over this lifetime, and
then only implementing HttpClient<'static> for HyperClient, allowing
other implementations to theoretically do the cool thing, but it
introduces so much line noise into the code that it's just not worth it.

Response body can at least still be efficient in that it implements
AsyncBufRead which lets you use the buffers the client allocated without
doing any extra copying, but sadly the request body requires a copy no
matter what.